updated open version

+++ commands/clearcommands.js
@@ -1,35 +1,37 @@
-const Discord = require("discord.js");
+let needDeleteLength;
 
 exports.run = (client, message, args) => {
-  if(message.member.roles.find(role => role.name === "Admin" || role.name === "Bot Controller"))
-  {
-    async function clear()
-      {
-        message.delete();
-        var needDelete = [];
-        message.channel.fetchMessages({limit: 100}).then(collected => { //collected is a Collection
-        collected.forEach(msg => {
-        if (msg.content.startsWith("~") || msg.content.startsWith("!") || msg.author.bot) needDelete.push(msg);
-        }
+	if(message.member.roles.find(role => role.name === 'Admin' || role.name === 'Bot Controller')) {
+		if (!args[0]) return message.reply('Set the number of messages you need to delete (1-100)');
+		const clearnumber = parseInt(args[0]);
+		if (clearnumber > 100) return message.reply('Set the number of messages you need to delete (1-100)');
+		message.channel.send('Clearing messages...');
+		clear(clearnumber).then;
+		setTimeout(() => {
+			message.channel.fetchMessages({ limit: 5 }).then(collected => {
+				collected.forEach(msg => {
+					if (msg.content.startsWith('Clearing')) msg.delete();
+				});
+			});
+			message.reply(`${needDeleteLength} Messages Cleared!`);
+		}, 3000);
 
-      );
-        message.channel.bulkDelete(needDelete);
-      });
-      }
-      message.channel.send("Clearing messages...");
-      clear();
+	}
+	else {
+		message.reply('You do not have a permission to run this command.');
+	}
 
-        setTimeout(() => {
-          message.channel.fetchMessages({limit: 5}).then(collected => { //collected is a Collection
-          collected.forEach(msg => {
-            if (msg.content.startsWith("Clearing")) msg.delete();
-          });
-        })
-          message.reply("Messages Cleared!");
-        }, 3000);
+	async function clear(clearnumber) {
+		message.delete();
+		const needDelete = [];
+		await message.channel.fetchMessages({ limit: clearnumber }).then(collected => {
+			collected.forEach(msg => {
+				if (msg.content.startsWith('~') || msg.content.startsWith('!') || msg.author.bot) needDelete.push(msg);
+			}
 
-  } else
-  {
-  message.reply("You do not have a permission to run this command.");
-  }
-}
+			);
+			needDeleteLength = needDelete.length;
+			message.channel.bulkDelete(needDelete);
+		});
+	}
+};
